+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### vNext
+
+- Show thing prefabHash and prefabName.
+- Show thing reference ids in slots.
+- Show thing health.
+- Show thing access state.
+- Support writing accessState to devices.
+- Items endpoint:
+  - item parent slotId and referenceId.
+  - Item quantity text.
+
 ### 1.1.0
 
 - Include string reference id for javascript consumption.